Implement logical deletion in the external obj level.
